Styles Update
Added preloaders to the site while loading the content from the
database. Lets the user know that it is working at least. Also addressed
the issue of the top nav showing up red prior to changing colors (might
be quick sometimes, but visible at slow speeds). This no longer occurs!